Add options --nobatch, --logdir, and --max_count.

+++ scripts/merge2.py
@@ -19,8 +19,12 @@
 # --stage <stage>     - Project stage.
 # --defname <defname> - Process files belonging to this definition (optional).
 # --database <path>   - Path of sqlite database file (default "merge.db").
+# --logdir <dir>      - Specify directory to store log files.
+#                       If specified each invocation generates a unique set of log files.
+#                       If not specified, output to stdout and stderr (no log files).
 # --max_size <bytes>  - Maximum merged file size in bytes (default 2.5e9).
 # --min_size <bytes>  - Minimum merged file size in bytes (default 1e9).
+# --max_count <n>     - Maximum number of files to merge per merged file (default no limit).
 # --max_age <seconds> - Maximum unmerged file age in seconds (default 72 hours).
 #                       Optionally use suffix 'h' for hours, 'd' for days.
 # --max_projects <n>  - Maximum number of projects.
@@ -30,6 +34,9 @@
 # --phase1            - Phase 1 (scan unmerged files).
 # --phase2            - Phase 2 (submit & monitor projects).
 # --phase3            - Phase 3 (monitor and clean up merged files).
+# --nobatch           - Inform the script that no merging batch jobs are
+#                       running or pending.  This sets various timeouts
+#                       related to sam projects to zero.
 #
 ######################################################################
 #
@@ -176,8 +183,9 @@ class MergeEngine:
     # Constructor.
 
     def __init__(self, xmlfile, projectname, stagename, defname,
-                 database, max_size, min_size, max_age, 
-                 max_projects, max_groups, query_limit, file_limit):
+                 database, max_size, min_size, max_count, max_age, 
+                 max_projects, max_groups, query_limit, file_limit,
+                 nobatch):
 
         # Open database connection.
 
@@ -229,11 +237,13 @@ class MergeEngine:
         self.defname = defname     # File selectionn dataset definition.
         self.max_size = max_size   # Maximum merge file size in bytes.
         self.min_size = min_size   # Minimum merge file size in bytes.
+        self.max_count = max_count # Maximum number of files to merge per merged file.
         self.max_age = max_age     # Maximum unmerged file age in seconds.
         self.max_projects = max_projects  # Maximum number of sam projects.
         self.max_groups = max_groups      # Maximum number of new merges groups per invocation.
         self.query_limit = query_limit    # Maximum number of files to query from sam.
         self.file_limit = file_limit      # Maximum number of unmerged files.
+        self.nobatch = nobatch     # Flag indicating that no batch jobs are pending.
 
         # Cache of directory contents.
 
@@ -389,6 +399,41 @@ CREATE TABLE IF NOT EXISTS unmerged_files (
                 self.dircache[dir] = set()
         if base in self.dircache[dir]:
             result = True
+        return result
+
+
+    # Get multiple metadata function.
+    # Similar as samweb.getMultipleMetadata, but no implicit maximum size.
+
+    def get_multiple_metadata(self, file_names):
+        result = []
+        q = []
+        print 'Getting multiple metadata for %d files.' % len(file_names)
+
+        # Loop over files.
+
+        for f in file_names:
+            #print 'Getting multiple metadata for file %s' %f
+            q.append(f)
+
+            # Maybe flush queue.
+
+            if len(q) >= self.metadata_queue_max:
+                mds = self.samweb.getMultipleMetadata(q)
+                for md in mds:
+                    result.append(md)
+                q = []
+
+        # Final queue flush.
+
+        if len(q) > 0:
+            mds = self.samweb.getMultipleMetadata(q)
+            for md in mds:
+                result.append(md)
+
+        # Done.
+
+        print 'Got metadata for %d files.' % len(result)
         return result
 
 
@@ -646,11 +691,11 @@ CREATE TABLE IF NOT EXISTS unmerged_files (
                 print 'File is already on tape.'
 
                 # File is on tape
-                # Modify metadata to set merge.merged flag to be true, so that this
+                # Modify metadata to set merge.merge flag to be false, so that this
                 # file will no longer be considered for merging.
 
-                mdmod = {'merge.merged': 1}
-                print 'Updating metadata to set merged flag.'
+                mdmod = {'merge.merge': 0}
+                print 'Updating metadata to reset merge flag.'
                 self.modifyFileMetadata(f, mdmod)
 
                 # Delete and remove any disk locations from sam.
@@ -679,7 +724,7 @@ CREATE TABLE IF NOT EXISTS unmerged_files (
                     if loc['location_type'] == 'disk':
                         dir = os.path.join(loc['mount_point'], loc['subdir'])
                         fp = os.path.join(dir, f)
-                        if self.exists(fp):
+                        if larbatch_posix.exists(fp):
                             print 'Location OK.'
                             on_disk = True
                         else:
@@ -728,28 +773,43 @@ CREATE TABLE IF NOT EXISTS unmerged_files (
         # Make sure add list is in form of a python set.
 
         add_files = set(flist)
-
-        # Make sure the number of files isn't too large for sqlite to handle.
-
-        while len(add_files) > 500:
-            add_files.pop()
-
         print '\n%d files in initial add group.' % len(add_files)
 
         # Query database to see which of these files already exist.
+        # Limit size of queries to what sqlite can handle.
 
         c = self.conn.cursor()
-        placeholders = ('?,'*len(add_files))[:-1]
-        q = 'SELECT name FROM unmerged_files WHERE name IN (%s);' % placeholders
-        c.execute(q, tuple(add_files))
-        rows = c.fetchall()
-        self.conn.commit()
+        uq = []
+        existing_files = []
+        for f in add_files:
+            uq.append(f)
 
-        for row in rows:
-            f = row[0]
+            # Maybe flush add queue.
+
+            if len(uq) >= 500:
+                placeholders = ('?,'*len(uq))[:-1]
+                q = 'SELECT name FROM unmerged_files WHERE name IN (%s);' % placeholders
+                c.execute(q, uq)
+                rows = c.fetchall()
+                for row in rows:
+                    existing_files.append(row[0])
+                uq = []
+
+        # Final queue flush.
+
+        if len(uq) > 0:
+            placeholders = ('?,'*len(uq))[:-1]
+            q = 'SELECT name FROM unmerged_files WHERE name IN (%s);' % placeholders
+            c.execute(q, uq)
+            rows = c.fetchall()
+            for row in rows:
+                existing_files.append(row[0])
+
+        # Remove existing files from add queue.
+
+        for f in existing_files:
             print 'Ignoring %s' % f
             add_files.discard(f)
-
         print '\b%d files in final add list.' % len(add_files)
 
         # Loop over files in add list and do bulk adds.
@@ -799,6 +859,24 @@ CREATE TABLE IF NOT EXISTS unmerged_files (
     def merge_group(self, md):
 
         group_id = -1
+
+        # Check that all seven required metadata fields are included.  If not return 0.
+        # Metadata field 'data_stream' is optional.
+
+        if not 'file_type' in md:
+            return 0
+        if not 'file_format' in md:
+            return 0
+        if not 'data_tier' in md:
+            return 0
+        if not 'ub_project.name' in md:
+            return 0
+        if not 'ub_project.stage' in md:
+            return 0
+        if not 'ub_project.version' in md:
+            return 0
+        if not 'runs' in md:
+            return 0
 
         # Create group 8-tuple.
 
@@ -959,7 +1037,7 @@ CREATE TABLE IF NOT EXISTS unmerged_files (
                 # Check for duplicate processed files in ths project.
 
                 parents = set()
-                mds = self.samweb.getMultipleMetadata(file_names)
+                mds = self.get_multiple_metadata(file_names)
                 for md in mds:
                     f = md['file_name']
                     if md.has_key('parents'):
@@ -1007,6 +1085,9 @@ CREATE TABLE IF NOT EXISTS unmerged_files (
                 if num_jobs > nfiles:
                     num_jobs = nfiles
                 max_files_per_job = (nfiles - 1) / num_jobs + 1
+                if max_files_per_job > self.max_count and self.max_count > 0:
+                    max_files_per_job = self.max_count
+                    num_jobs = (nfiles - 1) / max_files_per_job + 1
                 print 'Number of files = %d' % nfiles
                 print 'Number of batch jobs = %d' % num_jobs
                 print 'Maximum files per job = %d' % max_files_per_job
@@ -1052,7 +1133,7 @@ CREATE TABLE IF NOT EXISTS unmerged_files (
 
             # If start time is older than 24 hours, stop this project.
 
-            if dtsec > 24*3600:
+            if self.nobatch or dtsec > 24*3600:
                 result = True
 
         else:
@@ -1191,11 +1272,29 @@ CREATE TABLE IF NOT EXISTS unmerged_files (
                                             print 'Unmerged file %s' % consumed_file
 
                                         # Update process id join with unmerged file.
+                                        # Make sure query doesn't get too large for sqlite
+                                        # to handle.
 
-                                        placeholders = ('?,'*len(consumed_files))[:-1]
-                                        q = '''UPDATE unmerged_files SET sam_process_id=? 
-                                               WHERE name IN (%s);''' % placeholders
-                                        c.execute(q, (merge_id,) + tuple(consumed_files))
+                                        uq = []
+                                        for f in consumed_files:
+                                            uq.append(f)
+
+                                            # Maybe flush queue.
+
+                                            if len(uq) >= 500:
+                                                placeholders = ('?,'*len(uq))[:-1]
+                                                q = '''UPDATE unmerged_files SET sam_process_id=? 
+                                                       WHERE name IN (%s);''' % placeholders
+                                                c.execute(q, (merge_id,) + tuple(uq))
+                                                uq = []
+
+                                        # Final queue flush.
+
+                                        if len(uq) > 0:
+                                            placeholders = ('?,'*len(uq))[:-1]
+                                            q = '''UPDATE unmerged_files SET sam_process_id=? 
+                                                   WHERE name IN (%s);''' % placeholders
+                                            c.execute(q, (merge_id,) + tuple(uq))
                                         self.conn.commit()
 
                     # Update project status to 3.
@@ -1231,7 +1330,7 @@ CREATE TABLE IF NOT EXISTS unmerged_files (
                             dtsec = dt.total_seconds()
 
                             #if dtsec > 10800:
-                            if dtsec > 3600:
+                            if self.nobatch or dtsec > 3600:
 
                                 print 'Project ended: %s' % sam_project
                                 prj_ended = True
@@ -1308,7 +1407,7 @@ CREATE TABLE IF NOT EXISTS unmerged_files (
                         dt = now - stime
                         print 'Project age = %s' % dt
 
-                        if dt.total_seconds() > 24*3600:
+                        if self.nobatch or dt.total_seconds() > 24*3600:
 
                             # If project age is greater than 24 hours, start and then
                             # immediately stop this project, so that no batch job can
@@ -1850,11 +1949,27 @@ CREATE TABLE IF NOT EXISTS unmerged_files (
                     self.flush_metadata()
 
                     # Construct a query to do the database deletions in a single query.
+                    # Make sure that the query doesn't get too big for sqlite.
 
                     print 'Deleting unmerged files from database.'
-                    placeholders = ('?,' * len(unmerged_files))[:-1]
-                    q = 'DELETE FROM unmerged_files WHERE name IN (%s);' % placeholders
-                    c.execute(q, unmerged_files)
+                    uq = []
+                    for f in unmerged_files:
+                        uq.append(f)
+
+                        # Maybe flush queue.
+
+                        if len(uq) >= 500:
+                            placeholders = ('?,' * len(uq))[:-1]
+                            q = 'DELETE FROM unmerged_files WHERE name IN (%s);' % placeholders
+                            c.execute(q, uq)
+                            uq = []
+
+                    # Final queue flush.
+
+                    if len(uq) > 0:
+                        placeholders = ('?,' * len(uq))[:-1]
+                        q = 'DELETE FROM unmerged_files WHERE name IN (%s);' % placeholders
+                        c.execute(q, uq)
                     self.conn.commit()
 
                     # End of loop over unmerged files.
@@ -2003,8 +2118,10 @@ def main(argv):
     stagename = ''
     database = 'merge.db'
     defname = ''
+    logdir = ''
     max_size = 2500000000
     min_size = 1000000000
+    max_count = 0
     max_age = 3*24*3600
     max_projects = 500
     max_groups = 100
@@ -2013,6 +2130,7 @@ def main(argv):
     do_phase1 = False
     do_phase2 = False
     do_phase3 = False
+    nobatch = False
 
     args = argv[1:]
     while len(args) > 0:
@@ -2034,11 +2152,17 @@ def main(argv):
         elif args[0] == '--defname' and len(args) > 1:
             defname = args[1]
             del args[0:2]
+        elif args[0] == '--logdir' and len(args) > 1:
+            logdir = args[1]
+            del args[0:2]
         elif args[0] == '--max_size' and len(args) > 1:
             max_size = int(args[1])
             del args[0:2]
         elif args[0] == '--min_size' and len(args) > 1:
             min_size = int(args[1])
+            del args[0:2]
+        elif args[0] == '--max_count' and len(args) > 1:
+            max_count = int(args[1])
             del args[0:2]
         elif args[0] == '--max_age' and len(args) > 1:
             if args[1][-1] == 'h' or args[1][-1] == 'H':
@@ -2069,9 +2193,45 @@ def main(argv):
         elif args[0] == '--phase3':
             do_phase3 = True
             del args[0]
+        elif args[0] == '--nobatch':
+            nobatch = True
+            del args[0]
         else:
             print 'Unknown option %s' % args[0]
             return 1
+
+    # Check if we want to generate log files.
+
+    if logdir != '':
+
+        # Try to make logdir if it doesn't exist.
+
+        if not os.path.exists(logdir):
+            os.makedirs(logdir)
+
+        # Make sure log directory exists.
+            
+        if os.path.exists(logdir):
+
+            # Generate unique names for stdout and stderr log files using current time
+            # according to the pattern merge_YYYYmmDD_HHMM.out/.err
+
+            now = datetime.datetime.now()
+            merge_name = 'merge_%s' % datetime.datetime.strftime(now, '%Y%m%d_%H%M')
+            outpath = '%s/%s.out' % (logdir, merge_name)
+            errpath = '%s/%s.err' % (logdir, merge_name)
+
+            # Override sys.stdout and sys.stderr
+
+            sys.stdout = open(outpath, 'w')
+            sys.stderr = open(errpath, 'w')
+
+        else:
+
+            # If log directory doesn't exist, write output to stdout and stderr.
+
+            print 'Log directory does not exist.'
+            logdir = ''
 
     # If no phase option, do all three phases.
 
@@ -2083,8 +2243,9 @@ def main(argv):
     # Create merge engine.
 
     engine = MergeEngine(xmlfile, projectname, stagename, defname,
-                         database, max_size, min_size, max_age,
-                         max_projects, max_groups, query_limit, file_limit)
+                         database, max_size, min_size, max_count, max_age,
+                         max_projects, max_groups, query_limit, file_limit,
+                         nobatch)
     if do_phase1:
         engine.update_unmerged_files()
     if do_phase2:
